# Faster SHA256 OpenSSL hashing