--- a/.threading_canary
+++ b/.threading_canary
@@ -1,1 +1,1 @@
-THIS looks like a job for Threading Canar!y!1!..
+THIS looks like a job for Threading Canard!y!1!..