--- a/src/clients/perf/scenario_perf_client.cpp
+++ b/src/clients/perf/scenario_perf_client.cpp
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
+#include "crypto/openssl/hash.h"
 #include "ds/files.h"
 #include "perf_client.h"
 
@@ -131,6 +132,8 @@ public:
 int main(int argc, char** argv)
 {
   logger::config::default_init();
+  logger::config::level() = LoggerLevel::INFO;
+  crypto::openssl_sha256_init();
 
   CLI::App cli_app{"Scenario Perf Client"};
   ScenarioPerfClientOptions options(cli_app, argv[0]);