--- a/src/crypto/openssl/hash.h
+++ b/src/crypto/openssl/hash.h
@@ -78,4 +78,6 @@ namespace crypto
   };
 
   void openssl_sha256(const std::span<const uint8_t>& data, uint8_t* h);
+  void openssl_sha256_init();
+  void openssl_sha256_shutdown();
 }