--- a/src/enclave/enclave.h
+++ b/src/enclave/enclave.h
@@ -7,6 +7,7 @@
 #include "ccf/node_subsystem_interface.h"
 #include "ccf/pal/enclave.h"
 #include "ccf/pal/mem.h"
+#include "crypto/openssl/hash.h"
 #include "ds/oversized.h"
 #include "enclave_time.h"
 #include "indexing/enclave_lfs_access.h"
@@ -94,6 +95,7 @@ namespace ccf
     {
       ccf::pal::initialize_enclave();
       ccf::initialize_verifiers();
+      crypto::openssl_sha256_init();
 
       // From
       // https://software.intel.com/content/www/us/en/develop/articles/how-to-use-the-rdrand-engine-in-openssl-for-random-number-generation.html
@@ -195,6 +197,7 @@ namespace ccf
       LOG_TRACE_FMT("Shutting down enclave");
       ccf::shutdown_verifiers();
       ccf::pal::shutdown_enclave();
+      crypto::openssl_sha256_shutdown();
     }
 
     CreateNodeStatus create_new_node(
@@ -275,6 +278,7 @@ namespace ccf
 
     bool run_main()
     {
+      crypto::openssl_sha256_init();
       LOG_DEBUG_FMT("Running main thread");
 #ifndef VIRTUAL_ENCLAVE
       try
@@ -473,6 +477,8 @@ namespace ccf
         LOG_INFO_FMT("Enclave stopped successfully. Stopping host...");
         RINGBUFFER_WRITE_MESSAGE(AdminMessage::stopped, to_host);
 
+        crypto::openssl_sha256_shutdown();
+
         return true;
       }
 #ifndef VIRTUAL_ENCLAVE
@@ -483,6 +489,7 @@ namespace ccf
         // exceptions bubble up to here and cause the node to shutdown.
         RINGBUFFER_WRITE_MESSAGE(
           AdminMessage::fatal_error_msg, to_host, std::string(e.what()));
+        crypto::openssl_sha256_shutdown();
         return false;
       }
 #endif
@@ -500,6 +507,8 @@ namespace ccf
 
     bool run_worker()
     {
+      crypto::openssl_sha256_init();
+
       LOG_DEBUG_FMT("Running worker thread");
 #ifndef VIRTUAL_ENCLAVE
       try
@@ -511,12 +520,14 @@ namespace ccf
           msg->data.tid, std::move(msg));
 
         threading::ThreadMessaging::instance().run();
+        crypto::openssl_sha256_shutdown();
       }
 #ifndef VIRTUAL_ENCLAVE
       catch (const std::exception& e)
       {
         RINGBUFFER_WRITE_MESSAGE(
           AdminMessage::fatal_error_msg, to_host, std::string(e.what()));
+        crypto::openssl_sha256_shutdown();
         return false;
       }
 #endif